Work In Progress: A minimal std::expected<T, E>
This change includes a minimal implementation of the std::expected<T, E>
proposal in P0323R3 (http://wg21.link/P0323R3) which is hosted in the
`cppnetlib` repository.

In the process we also document some of our guidelines on the structure
of the repository, on naming files, directory structure, and dependency
requirements.

Still TODO:

* Fill out the test suite.
* Update documentation on using vcpkg on dependencies.
* Work our way up to a minimal interface for a connecton type, using our
  std::expected utility type.

This is related to the work in netlib/#5 for implementing a minimal
connection type suitable as a building block for higher level protocol
implementations.